update locations in db, move asyncForEach to backend/bin/
