+++ backend/models/location.js
@@ -21,6 +21,10 @@ const locationSchema = new Schema({
     type: Schema.Types.ObjectId,
     ref: 'locationTypes'
   } ],
+  locationTypeId: {
+    type: Schema.Types.ObjectId,
+    ref: 'locationTypes'
+  },
   description: {
     type: String,
     intl: true
